4-structured-outputs-pydantic/pyproject.toml
[project]
name = "4-structured-outputs-pydantic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "dotenv==0.9.9",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.179.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "httplib2==0.30.0",
    "idna==3.10",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "pypdf==6.0.0",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
]
